+++ test/5-getConfigSources.js
@@ -1,8 +1,9 @@
 
 
 // Dependencies
-var vows = require('vows'),
-    assert = require('assert');
+var vows   = require('vows'),
+    assert = require('assert'),
+    Path   = require('path');
 
 exports.DeferredTest = vows.describe('Tests config.util.getConfigSources').addBatch({
   'tests with NODE_CONFIG env set, and --NODE_CONFIG command line flag': {
@@ -70,7 +71,32 @@ exports.DeferredTest = vows.describe('Tests config.util.getConfigSources').addBa
     "The keys for each object are 'name', 'original', and 'parsed'": function(topic) {
         assert.deepEqual(Object.keys(topic[0]).sort(), ['name','original','parsed']);
     },
+ },
+
+
+ 'Files which return empty objects still end up in getConfigSources()': {
+    topic: function () {
+      // Change the configuration directory for testing
+      process.env.NODE_CONFIG_DIR = __dirname + '/5-config';
+
+      process.env.NODE_ENV='empty';
+      delete process.env.NODE_CONFIG;
+      process.argv = [];
+      var config = requireUncached('../lib/config');
+      return config.util.getConfigSources();
+    },
+
+    'Three files should result in 3 entries': function(topic) {
+        assert.equal(topic.length,3);
+    },
+
+    'Second file is named empty': function (topic) {
+      assert.equal(Path.basename(topic[1].name), 'empty.json');
+    },
+
  }
+
+
 
 });